harvester_api/tests/test_managers.py
from unittest import TestCase, mock
from json.decoder import JSONDecodeError

from harvester_api.api import HarvesterAPI
from harvester_api.managers import (DEFAULT_NAMESPACE, BaseManager,
                                    HostManager, ImageManager)


class TestBaseManager(TestCase):
    def setUp(self):
        self.api = mock.MagicMock(sepc=HarvesterAPI)
        self.mgr = BaseManager(self.api)
        self.API_VERSION = "TEST_API_VERSION"
        self.api.API_VERSION = self.API_VERSION

    def tearDown(self):
        self.api.reset_mock()

    def test_weakref(self):
        class FakeAPI:
            pass

        api = FakeAPI()
        bm = BaseManager(api)

        self.assertEqual(bm.api, api)

        with self.assertRaises(ReferenceError):
            del api
            bm.api

    def test__inject_data(self):
        raw_data = dict(some="{API_VERSION}",
                        thing=42,
                        api="unknown/{API_VERSION}")

        data = self.mgr._inject_data(raw_data)

        self.assertEqual(raw_data.keys(), data.keys())
        self.assertEqual(data['some'], self.api.API_VERSION)
        self.assertEqual(data['api'], "unknown/%s" % self.api.API_VERSION)

    def test__delegate(self):
        m_resp = mock.MagicMock()
        self.api._get.return_value = m_resp

        # Case 1: general case
        resp = self.mgr._delegate("_get", "/test/path")

        self.assertEqual(resp, (m_resp.status_code, m_resp.json.return_value))

        # Case 2: raw response
        m_resp.reset_mock()

        resp = self.mgr._delegate("_get", "/test/path", raw=True)

        self.assertEqual(resp, m_resp)

        # Case 3: json decode error
        m_resp.reset_mock()

        exception = JSONDecodeError("test Error", doc="", pos=42)
        m_resp.json.side_effect = exception
        resp = self.mgr._delegate("_get", "/test/path")

        self.assertEqual(resp, (m_resp.status_code,
                                dict(error=exception, response=m_resp)))

    def test__update(self):
        path, data = "/test/path", dict(test="data")

        # Case 1: default
        self.mgr._update(path, data)

        self.assertIn(path, self.api._put.call_args[0][0])
        self.assertDictEqual(dict(json=data), self.api._put.call_args[1])

        # Case 2: as_json = False
        self.mgr._update(path, data, as_json=False)

        self.assertDictEqual(dict(data=data), self.api._put.call_args[1])

        # Case 3: keyword argument overwrite
        self.mgr._update(path, data, json=dict())

        self.assertDictEqual(dict(json=data), self.api._put.call_args[1])


class TestHostManager(TestCase):
    def setUp(self):
        self.api = mock.MagicMock(sepc=HarvesterAPI)
        self.mgr = HostManager(self.api)

    def tearDown(self):
        self.api.reset_mock()

    def test_create(self):
        with self.assertRaises(NotImplementedError):
            self.mgr.create()

    def test_get(self):
        # Case 1: specific node
        node_name = "called"
        self.mgr.get(node_name)

        self.assertIn(node_name, self.api._get.call_args[0][0])

        # Case 2: all
        self.mgr.get()
        self.assertTrue(self.api._get.call_args[0][0].endswith("/"))

    def test_delete(self):
        node_name = "called"
        self.mgr.delete(node_name)

        self.assertIn(node_name, self.api._delete.call_args[0][0])

    def test_update(self):
        node_name = "called"
        data = dict(test="data")

        self.mgr.update(node_name, data)

        self.assertIn(node_name, self.api._put.call_args[0][0])
        self.assertDictEqual(dict(json=data), self.api._put.call_args[1])

    def test_get_metrics(self):
        # Case 1: specific node
        node_name = "called"
        self.mgr.get_metrics(node_name)

        self.assertIn(node_name, self.api._get.call_args[0][0])

        # Case 2: all
        self.mgr.get_metrics()

        self.assertTrue(self.api._get.call_args[0][0].endswith("/"))


class TestImageManager(TestCase):
    def setUp(self):
        self.api = mock.MagicMock(sepc=HarvesterAPI)
        self.mgr = ImageManager(self.api)
        self.API_VERSION = "TEST_API_VERSION"
        self.api.API_VERSION = self.API_VERSION

    def tearDown(self):
        self.api.reset_mock()

    def test_get(self):
        name, namespace = "image", "the-namespace"

        # Case 1: default
        self.mgr.get(name)

        self.assertIn(name, self.api._get.call_args[0][0])
        self.assertIn(DEFAULT_NAMESPACE, self.api._get.call_args[0][0])

        # Case 2: specific namespace
        self.mgr.get(name, namespace)

        self.assertIn(name, self.api._get.call_args[0][0])
        self.assertIn(namespace, self.api._get.call_args[0][0])

    def test_create_data(self):
        name, url, desc, stype = "name", "url", "desc", "stype"
        namespace, display_name = "namespace", "displayName"

        # Case 1: display name not assigned
        data = self.mgr.create_data(name, url, desc, stype, namespace)

        self.assertEqual(name, data['metadata']['name'])
        self.assertEqual(url, data['spec']['url'])
        self.assertEqual(desc, data['metadata']['annotations']['field.cattle.io/description'])
        self.assertEqual(stype, data['spec']['sourceType'])
        self.assertEqual(namespace, data['metadata']['namespace'])
        self.assertEqual(name, data['spec']['displayName'])
        self.assertEqual(self.API_VERSION, data['apiVersion'])

        # Case 2: display name assigned

        data = self.mgr.create_data(name, url, desc, stype, namespace, display_name)

        self.assertEqual(display_name, data['spec']['displayName'])

    def test_create(self):
        name, url, namespace = "ImageName", "testURL", "TestNamespace"

        # Case 1: default namespace
        self.mgr.create(name, url)

        self.assertNotIn(name, self.api._post.call_args[0][0])
        self.assertIn(DEFAULT_NAMESPACE, self.api._post.call_args[0][0])

        # Case 2: specific namespace
        self.mgr.create(name, url, namespace=namespace)

        self.assertIn(namespace, self.api._post.call_args[0][0])

    def test_update(self):
        name, namespace = "TestImageName", "TestNamespace"
        data = dict(metadata=dict(namespace=namespace))

        # Case 1: namespace miss
        self.mgr.update(name, dict())

        self.assertIn(name, self.api._put.call_args[0][0])
        self.assertNotIn(namespace, self.api._put.call_args[0][0])

        # Case 2: specific namespace
        self.mgr.update(name, data)

        self.assertIn(name, self.api._put.call_args[0][0])
        self.assertIn(namespace, self.api._put.call_args[0][0])

    def test_delete(self):
        name, namespace = "TestImageName", "TestNamespace"

        self.mgr.delete(name, namespace)

        self.assertIn(name, self.api._delete.call_args[0][0])
        self.assertIn(namespace, self.api._delete.call_args[0][0])
